refactor: permission checks for update and delete actions in tweets and videos

+++ src/controllers/tweet.controller.js
@@ -63,6 +63,10 @@ const updateTweet = asyncHandler(async (req, res) => {
     throw new ApiError(404, "Tweet does not exist");
   }
 
+  if (tweet.owner.toString() !== req.user?._id.toString()) {
+    throw new ApiError(403, "Forbidden request");
+  }
+
   const updatedTweet = await Tweet.findByIdAndUpdate(
     tweetId,
     {
@@ -91,6 +95,10 @@ const deleteTweet = asyncHandler(async (req, res) => {
   const tweet = await Tweet.findById(tweetId);
   if (!tweet) {
     throw new ApiError(404, "Tweet does not exist");
+  }
+
+  if (tweet.owner.toString() !== req.user?._id.toString()) {
+    throw new ApiError(403, "Forbidden request");
   }
 
   await Tweet.findByIdAndDelete(tweetId);
